Move inode-related types/methods to inode.rs
